Fix issue with the size of RECOOUT.dt
In the previous version, the size of the output time-difference array
(ro.dt) was fixed to 32 elements in length, presumablye because
that was all that was needed for TB.
But, this is much to short for the deep ARA stations which can have
at most 16 channels in use for a reconstruction, and therefore
120 active delays.

As it was written with only 32 delays, if you loaded >7 antennas
into the system of equations, when the reconstrucor is called
(Reco->doSphericalFit()), and it goes to evaluate ice->getTD(),
whic calls ice->get(), it ends up trying to access an element
of ro.dt[i] thta doesn't exist.

I have fixed this temporarily by increasing the size of ro.dt
but of course even that will break down if the stations get
more complex later. But it's good enough for now.