Add new AraRecoHandler class
Introduce a new class to the AraVertex library, the "AraRecoHandler"

The point of this class is to manage *inputs* to a reconstruction
tool like AraVertex. The base version of this class as it stands
comes with an SNR calculator (according to the UW definition)
and comes with a tool that will take an event, determine hit times
and give them to the AraVertex pointer a user passes to it.