AraVertex dev1 (#14)
* Bring AraVertex library into service

This commit mostly makes changes to CMakeLists.txt and LinkDef.h
necessary for AraRoot to actually compile AraVertex into a proper library
which can be compiled against.
For the moment, L2 functionality has been disabled (ARA never really "got into" L2)
but it would be good to restore the functionality long term.

* Fix issue with the size of RECOOUT.dt

In the previous version, the size of the output time-difference array
(ro.dt) was fixed to 32 elements in length, presumablye because
that was all that was needed for TB.
But, this is much to short for the deep ARA stations which can have
at most 16 channels in use for a reconstruction, and therefore
120 active delays.

As it was written with only 32 delays, if you loaded >7 antennas
into the system of equations, when the reconstrucor is called
(Reco->doSphericalFit()), and it goes to evaluate ice->getTD(),
whic calls ice->get(), it ends up trying to access an element
of ro.dt[i] thta doesn't exist.

I have fixed this temporarily by increasing the size of ro.dt
but of course even that will break down if the stations get
more complex later. But it's good enough for now.

* Add new AraRecoHandler class

Introduce a new class to the AraVertex library, the "AraRecoHandler"

The point of this class is to manage *inputs* to a reconstruction
tool like AraVertex. The base version of this class as it stands
comes with an SNR calculator (according to the UW definition)
and comes with a tool that will take an event, determine hit times
and give them to the AraVertex pointer a user passes to it.

* identation should be done with spaces, oops

* add example of how to use AraVertex class