Update AraRecoHandler to take antenna positions instead of AraGeomTool
This commit changes the AraRecoHandler hit finding function to take an array
of antenna positions instead of an araGeomTool. This is actually more efficient
so that the antenna locations do not need to be called over and over again.
There is an additional helper function added to populate the vector that's needed.