Add iostream to AraHkPlotter
iostream was apparently missing from AraHkPlotter includes. a new compilation (with ROOT 6.22.00 and gcc 4.5.8, not sure if they are connected in anyway identified the error). The error was: 'cerr' is not a member of 'std'. Adding iostream seems to fix it. Not sure why this wasn't breaking before now, but is fixed now.