Allocate atri pedestal arrays on the heap
The atri pedestal buffer array was allocated on the stack.
But the size of the array exceeded the default stack
size on some kernels, in particular we saw this happening on the cobalt nodes.
Which leads to a very hard to diagnose sefault.
So, instead, allocate the buffers with vectors so they go onto the heap.