make update to the RTC
Major update to the RayTraceCorrelator.
Namely, I have rewritten the getCorrFunctions method to use unique pointers
for internal management, and to return a vector of TGraphs, instead of TGraph*.
So std::vector<TGraph> instead of std::vector<TGraph*>.

I also modified the GetInterferometricMap routine to take as an input
the vector<TGraph>, and it now returns a TH2D, not TH2D*.
This is done to make these objects much more memory safe, especially
when called from pyroot.